\name{OneHandClapping-package}
\alias{OneHandClapping-package}
\alias{OneHandClapping}
\docType{package}
\title{
Prediction of condition-specific transcription factor interactions
}
\description{
Predictions of condition-specific TF interactions using gene activity data,
    such as mRNA expression measurements and a TF-target graph.
}
\details{
\tabular{ll}{
Package: \tab OneHandClapping\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-01-17\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 2)\cr
Imports: \tab methods, graphics\cr
Built: \tab R 2.14.1; ; 2012-01-18 14:19:32 UTC; unix\cr
}

Index:
\preformatted{
biogridList.71          TF interactions from BioGRID
salt.folds              mRNA folds after 30 min of 0.8M NaCl
                        stimulation relative to wildtype
yeastract               Yeastract database, TF-target relations
}

The main function is \code{\link{siscreen}}

}
\author{
Sebastian Dümcke

Maintainer: Sebastian Dümcke <duemcke@genzentrum.lmu.de>
}
\keyword{ package }
\seealso{
\code{\link{siscreen}}
}
