\name{salt.folds}
\alias{salt.folds}
\docType{data}
\title{mRNA folds after 30 min of 0.8M NaCl stimulation relative to wildtype}
\description{

Total mRNA measurements in Yeast BY4741 30 minutes after addition of 0.8M NaCl to the medium using the labelling approach
described in \cite{Miller et al. (MSB, 2010)}
}

\usage{data(salt.folds)}
\format{
named vector with folds. Names follow official Yeast nomenclature
}
\source{
Miller \emph{et al.} paper: 10.1038/msb.2010.112 (DOI)
}
\examples{
data(salt.folds)
head(salt.folds)
}
\keyword{datasets}
